# bin/main/de/oth/othappserver/api/RoomFinderApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.1.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package de.oth.othappserver.api

import de.oth.othappserver.model.FreeRoomBuilding
import de.oth.othappserver.model.RoomType
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import javax.validation.Valid
import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size

import kotlin.collections.List
import kotlin.collections.Map

@Validated
@RequestMapping("\${api.base-path:}")
interface RoomFinderApi {


    @GetMapping(
            value = ["/v1/freerooms"],
            produces = ["application/json"]
    )
    fun freeRooms(@Pattern(regexp="^((\\d\\d:\\d\\d)(:\\d\\d)?)|(now)$")  @RequestParam(value = "start", required = false) start: kotlin.String?
, @RequestParam(value = "date", required = false) date: java.time.LocalDate?
,@Pattern(regexp="^\\d\\d:\\d\\d$")  @RequestParam(value = "end", required = false) end: kotlin.String?
, @RequestParam(value = "building", required = false) building: kotlin.String?
, @RequestParam(value = "equipment", required = false) equipment: kotlin.collections.List<kotlin.Int>?
, @RequestParam(value = "roomType", required = false) roomType: kotlin.collections.List<RoomType>?
): ResponseEntity<List<FreeRoomBuilding>> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }
}