# bin/main/de/oth/othappserver/api/SemesterApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.1.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package de.oth.othappserver.api

import de.oth.othappserver.model.Error
import de.oth.othappserver.model.Lecture
import de.oth.othappserver.model.TimetableItem
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import javax.validation.Valid
import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size

import kotlin.collections.List
import kotlin.collections.Map

@Validated
@RequestMapping("\${api.base-path:}")
interface SemesterApi {


    @GetMapping(
            value = ["/v1/semester/{id}/lectures"],
            produces = ["application/json"]
    )
    fun semesterLectures( @PathVariable("id") id: kotlin.Int
, @RequestParam(value = "startDate", required = false) startDate: java.time.LocalDate?
, @RequestParam(value = "endDate", required = false) endDate: java.time.LocalDate?
): ResponseEntity<List<Lecture>> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }


    @GetMapping(
            value = ["/v1/semester/{id}/timetable"],
            produces = ["application/json"]
    )
    fun semesterTimetable( @PathVariable("id") id: kotlin.Int
, @RequestParam(value = "startDate", required = false) startDate: java.time.LocalDate?
, @RequestParam(value = "endDate", required = false) endDate: java.time.LocalDate?
): ResponseEntity<List<TimetableItem>> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }
}