# build/resources/main/static/openapi.yaml
openapi: 3.0.3
info:
  title: OTH-App API
  version: 'dev.20221124'
  contact:
    name: Studierendenvertretung
    url: https://stuv.othr.de
    email: studierendenvertretung@oth-regensburg.de
  x-logo:
    url: othr_logo.png
...
paths:
...
  "/v1/faculty/{identifier}":
    get:
      tags:
        - Faculty
      summary: Find Faculty by identifier
      description: Retrieve a single faculty filtered by the given identifier.
      operationId: faculty
      parameters:
        - name: identifier
          in: path
          description: Identifier of the faculty
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Faculty found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacultyDetail"
        '404':
          description: Faculty not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                404Error:
                  $ref: "#/components/examples/404Error"
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X GET 'localhost:8080/v1/faculty/ANK' -H 'Accept: application/json'
...
  "/v1/semester/{id}/lectures":
    get:
      tags:
        - Semester
      summary: Get Lectures
      description: |
        Gets all lectures of a semester.

        Returns only lecture IDs. See the <a href="/#operation/info">university information endpoint</a> to get complete lecture data.

        ### Further notes
        This data is computed by extracting each lecture from the current semesters timetable and **might be incomplete**.
      operationId: semesterLectures
      parameters:
        - name: id
          in: path
          description: Id of a semester
          required: true
          schema:
            type: integer
            example: 19578
        - name: startDate
          in: query
          description: Filter Lectures that take place after this start date. If left out, the current semesters start date will be used.
          required: false
          schema:
            type: string
            format: date
            example: "2022-10-16"
        - name: endDate
          in: query
          description: Filter Lectures that take place before this end date. If left out, the current semesters end date will be used.
          required: false
          schema:
            type: string
            format: date
            example: "2022-11-16"
      responses:
        '200':
          description: Semester lectures retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lecture"
        '404':
          description: Semester not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                404Error:
                  $ref: "#/components/examples/404Error"
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X GET 'https://im-vm-031.othr.de/v1/semester/7840/lectures' -H 'Accept: application/json'
  "/v1/semester/{id}/timetable":
    get:
      tags:
        - Semester
      summary: Get Timetable
      description: |
        Gets the timetable for a semester in the given time period. If either `startDate` or `endDate` are missing the current semesters start and end date are used.

        Result is ordered by `startDateTime`.
      operationId: semesterTimetable
      parameters:
        - name: id
          in: path
          description: Id of a semester
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          description: Start date for the requested timetable.
          required: false
          schema:
            type: string
            format: date
            example: "2022-10-16"
        - name: endDate
          in: query
          description: End date for the requested timetable.
          required: false
          schema:
            type: string
            format: date
            example: "2022-11-16"
      responses:
        '200':
          description: Semester lectures retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableItem"
        '404':
          description: Semester not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                404Error:
                  $ref: "#/components/examples/404Error"
      x-codeSamples:
        - lang: 'Shell'
          label: Default
          source: |
            # Timetable for Semester MIN (7840) for the current semesters time period.
            curl -X GET 'https://im-vm-031.othr.de/v1/semester/7840/timetable' \
                -H 'Accept: application/json'
        - lang: 'Shell'
          label: Custom date
          source: |
            # Timetable for Semester MIN (7840) for a custom time period.
            curl -X GET 'https://im-vm-031.othr.de/v1/semester/7840/timetable?startDate=2022-10-16&endDate=2022-11-16' \
                -H 'Accept: application/json'
...
  "/v1/freerooms":
    get:
      tags:
        - RoomFinder
      summary: List free rooms
      description: |
        Gets all free rooms for a given filter.

        For possible filter values for `equipment` and `building` see the `roomFinderFilter` attribute of the <a href="/#operation/info">university information endpoint</a> result.
      operationId: freeRooms
      parameters:
        - in: query
          name: start
          schema:
            type: string
            pattern: "^((\\d\\d:\\d\\d)(:\\d\\d)?)|(now)$"
            example:
              - "08:00"
              - "now"
          description: Start time from which free rooms are to be searched. Can either be a time like `08:00` for a concrete time or `now` for the current time.
        - in: query
          name: date
          schema:
            type: string
            format: date
            example: "2022-11-27"
          description: Date from which free rooms are to be searched. Needs to be between the current date and the next 7 days. If `null` the current date is used.
        - in: query
          name: end
          schema:
            type: string
            pattern: "^\\d\\d:\\d\\d$"
            example: "20:00"
          description: End time until which free rooms should be searched.
        - in: query
          name: building
          schema:
            type: string
            example: "K"
          description: Letter of the building to be searched. If omited, rooms in every building are searched.
        - in: query
          name: equipment
          schema:
            type: array
            example: 1
            items:
              type: integer
          description: Ids of equipment to filter. A room has to have one of the requested equipment to be returned.
        - in: query
          name: roomType
          schema:
            type: array
            items:
              $ref: "#/components/schemas/RoomType"
          description: Types of room to be searched for.
      responses:
        '200':
          description: Buildings with free rooms retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreeRoomBuilding"
      x-code-samples:
        - lang: 'curl'
          source: |
            curl -X GET 'https://im-vm-031.othr.de/v1/freerooms?start=now&building=K&date=2022-10-22&roomType=Hoersaal,CIP%20Pool' -H 'Accept: application/json'
  "/v1/feedback":
    post:
      tags:
        - Feedback
      summary: Provide Feedback
      description: Posts a new feedback entry (sent via email to a mailing list)
      operationId: feedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackItem"
      responses:
        '202':
          description: accepted
        '500':
          description: other Exception
        '504':
          description: IOException
  "/v1/candidates":
    get:
      tags:
        - Election
      summary: "Returns all candidates for the latest election."
      responses:
        '200':
          description: "All available candidates."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
      operationId: getCandidates
    post:
      tags:
        - Election
      summary: "Create a new candidate."
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CandidatePayload"
      responses:
        '200':
          description: "Created"
          content:
            text/plain:
              schema:
                type: string
                description: the id of the created candidate
                example: 4
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: createCandidate
      security:
        - ApiKeyAuth: []
  "/v1/candidates/{candidateId}":
    get:
      parameters:
        - name: candidateId
          in: path
          required: true
          description: "The id of the candidate to retrieve."
          schema:
            type: integer
      tags:
        - Election
      summary: "Returns a candidate of an election"
      responses:
        '200':
          description: "A candidate object."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getCandidateById
    put:
      parameters:
        - name: candidateId
          in: path
          required: true
          description: "The id of the candidate to update."
          schema:
            type: integer
      tags:
        - Election
      summary: "Update an existing candidate."
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Candidate"
      responses:
        '200':
          description: "Candidate was updated."
        '304':
          $ref: '#/components/responses/NotModified'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: updateCandidate
      security:
        - ApiKeyAuth: []
    delete:
      parameters:
        - name: candidateId
          in: path
          required: true
          description: "The id of the candidate to retrieve."
          schema:
            type: integer
      tags:
        - Election
      summary: "Delete a candidate."
      responses:
        '200':
          description: "Candidate deleted."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: deleteCandidate
      security:
        - ApiKeyAuth: []
  "/v1/candidates/{candidateId}/image":
    get:
      parameters:
        - name: candidateId
          in: path
          required: true
          description: "The id of the candidate."
          schema:
            type: integer
      tags:
        - Election
      summary: "Returns the image of a candidate."
      responses:
        '200': # Response
          description: OK
          content: # Response body
            image/*: # Media type
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getImageOfCandidate
    post:
      parameters:
        - name: candidateId
          in: path
          required: true
          description: "The id of the candidate."
          schema:
            type: integer
      tags:
        - Election
      summary: "Updates the image of a candidate"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: "Created"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: uploadPictureOfCandidate
      security:
        - ApiKeyAuth: []
  "/v1/election":
    get:
      tags:
        - Election
      summary: "Returns the election information of the latest election."
      responses:
        '200':
          description: "A candidate object."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectionInfo"
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: getElectionInfo
    post:
      tags:
        - Election
      summary: "Creates or updates the election information."
      description: Contains the necessary data for an election. Newlines and '\\n' can be used and will be normalized to '\\n' (no '\\r')
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ElectionInfo"
      responses:
        '200':
          description: "Created or updated new election information."
        '401':
          $ref: '#/components/responses/Unauthorized'
      operationId: createElectionInfo
      security:
        - ApiKeyAuth: []
    delete:
      tags:
        - Election
      summary: "Delete the latest election information."
      responses:
        '200':
          description: "Deleted latest election information."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      operationId: deleteElectionInfo
      security:
        - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: X-API-KEY  # name of the header, query parameter or cookie
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotModified:
      description: The specified resource was not modified, because nothing changed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized, specify an api token in the header 'X-API-KEY'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: RFC 7807 compliance error.
      properties:
        type:
          type: string
          example: https://othr.de/error/event/InvalidSyncToken
          description: A URI reference that identifies the problem type.
        title:
          type: string
          example: Invalid deltaToken.
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          example: 400
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
        detail:
          type: string
          example: The provided deltaToken is not valid anymore.
          description: A human-readable explanation specific to this occurrence of the problem.
...
    FacultyDetail:
      properties:
        identifier:
          type: string
          example: "IM"
          description: A short representation of the faculty's name. Uniquely identifies a faculty.
          minLength: 1
          maxLength: 3
        name:
          example: "Informatik und Mathematik"
          type: string
          minLength: 1
          description: The complete faculty's name.
        color:
          type: string
          example: "4A78D3"
          pattern: "[\\dA-F]{6}"
          maxLength: 6
          nullable: true
          description: A faculty's color in hex.
        feeds:
          type: array
          description: The Faculty's RSS feeds.
          items:
            $ref: "#/components/schemas/Feed"
        semesters:
          type: array
          description: The Faculty's Semesters.
          items:
            $ref: "#/components/schemas/Semester"
        lecturer:
          type: array
          description: The Faculty's Lecturers.
          items:
            $ref: "#/components/schemas/Lecturer"
      required:
        - identifier
        - name
        - semesters
        - lecturer
        - feeds
        - semesters
    Feed:
      type: object
      properties:
        rss:
          type: string
          description: RSS feed URI
          example: "https://www.oth-regensburg.de/fakultaeten/informatik-und-mathematik/schwarzes-brett/rss.xml"
        type:
          type: string
          description: Type of the RSS feed.
          enum:
            - "BULLETIN_BOARD"
      required:
        - rss
        - type
      description: A faculty's RSS Feed.
    FeedbackItem:
      type: object
      properties:
        category:
          type: string
          enum: [ Bug, Feedback, Idee ]
        message:
          type: string
        contactAddress:
          type: string
        images:
          type: array
          minItems: 0
          maxItems: 3
          items:
            type: string
            format: base64
            description: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAJYAAACWBAMAAADOL2zRAAAAG1BMVEXMzMyWlpaqqqq3t7fFxcW+vr6xsbGjo6OcnJyLKnDGAAAACXBIWXMAAA7EAAAOxAGVKw4bAAABAElEQVRoge3SMW+DMBiE4YsxJqMJtHOTITPeOsLQnaodGImEUMZEkZhRUqn92f0MaTubtfeMh/QGHANEREREREREREREtIJJ0xbH299kp8l8FaGtLdTQ19HjofxZlJ0m1+eBKZcikd9PWtXC5DoDotRO04B9YOvFIXmXLy2jEbiqE6Df7DTleA5socLqvEFVxtJyrpZFWz/pHM2CVte0lS8g2eDe6prOyqPglhzROL+Xye4tmT4WvRcQ2/m81p+/rdguOi8Hc5L/8Qk4vhZzy08DduGt9eVQyP2qoTM1zi0/uf4hvBWf5c77e69Gf798y08L7j0RERERERERERH9P99ZpSVRivB/rgAAAABJRU5ErkJggg=="
        additionalInfo:
          type: object
          additionalProperties:
            type: string
      required:
        - category
        - message
...
    Lecture:
      type: object
      properties:
        id:
          type: integer
          description: Id of the resource
          example: 17815
        name:
          type: string
          example: "Moderne Softwaretechniken"
          description: Name of the Lecture.
        shortName:
          type: string
          example: "MST"
          description: Short name of the Lecture.
      required:
        - id
        - name
        - shortName
    Semester:
      type: object
      properties:
        id:
          type: integer
          description: Id of the resource
          example: 7840
        name:
          type: string
          description: Short name of the Semester
          example: "MIN"
        longName:
          type: string
          description: Long name of the Semester
          example: "Master MIN"
      required:
        - id
        - name
        - longName
...
    TimetableItem:
      type: object
      properties:
        id:
          type: integer
          description: Id of the timetable item.
          example: 72543
        startDateTime:
          type: string
          format: 'date-time'
          description: Start date time of the timetable occurrence.
          example: "2022-11-16T15:15:00+01:00"
        endDateTime:
          type: string
          format: 'date-time'
          description: Start date time of the timetable occurrence.
          example: "2022-11-16T16:45:00+01:00"
        rooms:
          type: array
          items:
            type: integer
            description: Ids of the rooms where the lecture is located at.
            example: 534
        lecture:
          type: integer
          description: Id of the lecture.
          example: 8726
          nullable: true
      required:
        - id
        - startDateTime
        - endDateTime
        - rooms
...
    RoomType:
      type: string
      enum:
        - Hoersaal
        - CIP Pool
...
    FreeRoomBuilding:
      type: object
      properties:
        letter:
          type: string
          minLength: 1
          maxLength: 1
          example: "K"
          description: "Letter of the building."
        faculties:
          type: array
          description: "Name of the building."
          minItems: 1
          items:
            type: string
        rooms:
          type: array
          minItems: 0
          items:
            $ref: "#/components/schemas/FreeRoom"
          description: "A list of available and free rooms."
        plans:
          type: array
          items:
            type: object
            properties:
              floor:
                type: integer
                description: The buildings floor
                example: 0
              plan:
                type: string
                description: "URL pointing to a plan for the corresponding floor of the building."
                example: "https://www.oth-regensburg.de/fileadmin/media/lageplaene/Galgenbergstr_32_K_EG_.pdf"
      required:
        - letter
        - faculties
        - rooms
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
          description: "Identification number of the candidate"
          readOnly: true
        firstName:
          type: string
          example: "Max"
          description: "First name of the candidate."
        lastName:
          type: string
          example: "Mustermann"
          description: "Last name of the candidate."
        course:
          type: string
          example: "Informatik"
          description: "Course of studies the candidate applied to."
        term:
          type: integer
          example: 5
          description: "The term the candidate is currently graduating."
        offices:
          type: array
          description: "A list of offices the candidate is running for."
          minItems: 1
          items:
            type: string
            enum:
              - "Studentischer Konvent"
              - "Senat"
              - "Fakultaetsrat"
        goals:
          type: array
          description: "The goals the candidate wants to achieve during his/her time in office."
          items:
            type: string
        motto:
          type: string
          description: "The motto of the candidate."
          example: "Make OTH great again"
        faculty:
          type: string
          description: "The faculty of the candidate and his course."
          enum:
            - "ANK"
            - "ARC"
            - "ASG"
            - "BAU"
            - "BW"
            - "EI"
            - "IM"
            - "MB"
          example: "IM"
      required:
        - firstName
        - lastName
        - course
        - term
        - offices
        - goals
        - faculty
    CandidatePayload:
      allOf:
        - $ref: "#/components/schemas/Candidate"
        - type: object
          properties:
            file:
              description: "Image of the candidate"
              type: string
              format: binary
      required:
        - firstName
        - lastName
        - course
        - term
        - offices
        - goals
        - faculty
    ElectionInfo:
      type: object
      properties:
        electionStart:
          type: string
          format: date
          example: "2022-11-13"
          description: "Date of the start of the election."
        electionEnd:
          type: string
          format: date
          example: "2022-11-16"
          description: "Date of the end of the election."
        startShowElectionInfo:
          type: string
          format: date
          example: "2022-11-13"
          description: "Date when to show the election info in the app."
        endShowElectionInfo:
          type: string
          format: date
          example: "2022-11-16"
          description: "Date when to stop showing the election info."
        howToVoteShort:
          type: object
          example: "Go somewhere and make your cross."
          description: "Short description of how to vote."
        howToVoteLong:
          type: object
          example: "Go somewhere and make your cross, described in detail."
          description: "Detailed description of how to vote."
        whatToVoteSenat:
          type: object
          example: "You can vote someone for senate."
          description: "Description of voting for the senate."
        whatToVoteKonvent:
          type: object
          example: "You can vote someone for the Studentischen Konvent."
          description: "Description of voting for the Studentischen Konvent."
        whatToVoteFaculty:
          type: object
          example: "You can vote someone for the Fakultätsrat."
          description: "Description of voting for the Fakultätsrat."
      required:
        - electionStart
        - electionEnd
        - startShowElectionInfo
        - endShowElectionInfo
  examples:
...
    404Error:
      description: A generic 404 Not Found error.
      value:
        title: Not Found.
        status: 404
        detail: Requested resource not found.